+++ .eslintrc.js
@@ -1,6 +1,6 @@
 module.exports = {
     root: true,
-    parser: 'babel-eslint',
+    parser: '@babel/eslint-parser',
     parserOptions: {
         sourceType: 'module',
     },
@@ -10,9 +10,11 @@ module.exports = {
     extends: ['standard', 'prettier'],
     plugins: ['prettier'],
     rules: {
-        'prettier/prettier': 'error',
+        'prettier/prettier': ['warn', {
+            'singleQuote': true
+        }],
         'arrow-parens': 0,
         'generator-star-spacing': 0,
-        'no-debugger': process.env.NODE_ENV === 'production' ? 2 : 0,
+        'no-debugger': process.env.NODE_ENV === 'production' ? 2 : 0
     },
 };