+++ src/core/attrs.js
@@ -1,0 +1,10 @@
+const attrsDefault = {
+    toggle: 'easy-toggle',
+    add: 'easy-add',
+    remove: 'easy-remove',
+    class: 'easy-class',
+    rcoe: 'easy-rcoe',
+    parallel: 'easy-parallel',
+}
+
+export default attrsDefault;