+++ src/easy-toggler.umd.js
@@ -1,0 +1,6 @@
+import easyTogglerHandler from './core';
+import attrsDefault from './core/attrs';
+
+document.addEventListener('DOMContentLoaded', () => {
+    document.addEventListener('click', (e) => easyTogglerHandler(e, attrsDefault));
+});