+++ CHANGELOG.md
@@ -1,20 +1,29 @@
 # CHANGELOG
 
-### 1.3.0
+## 2.0.0 (2022-06-16)
+**Closed issues:**
+- fix "Don't work remove class after added data-easy-parallel #3"
+
+**Implemented enhancements:**
+- code refactoring
+- add `easy-add`
+- add `easy-remove`
+
+## 1.3.0
 **Date: Jun 13, 2022**
 - Add base test
 - Update plugin structure (add jest, eslint, babel and more...)
 
 Thanks to everyone who wrote about the plugin: https://www.youtube.com/watch?v=Ov5wG-MaDzk ❤
 
-### 1.2.0
+## 1.2.0
 **Date: Aug 1, 2020**
 Added attribute "data-easy-parallel"
 
-### 1.0.0
+## 1.0.1
 **Date: Aug 1, 2020**
 V 1.0.0 with fixes
 
-### 1.0.0
+## 1.0.0
 **Date: Jul 31, 2020**
 Lite version with base functional